pyproject.toml
[project]
name = "fragmenter"
version = "0.1.0"
description = "Local-first RAG indexer for repos, docs, and PDFs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Karl Lundgren", email = "k.github@lundgrens.net" }]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "llm",
    "vector-store",
    "llamaindex",
    "chromadb",
    "cli",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "chromadb>=0.4.0",
    "llama-index>=0.14.9",
    "llama-index-vector-stores-chroma>=0.1.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
    "fake-useragent>=1.5.1",
    "pypdf>=6.4.0",
    "tree-sitter>=0.21.3",
    "tree-sitter-language-pack>=0.13.0",
    "pydantic-settings>=2.12.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
openai = ["llama-index-llms-openai>=0.3.0"]
anthropic = ["llama-index-llms-anthropic>=0.3.0"]
ollama = ["llama-index-llms-ollama>=0.3.0"]
huggingface = ["llama-index-llms-huggingface>=0.3.0"]
all-providers = [
    "fragmenter[openai]",
    "fragmenter[anthropic]",
    "fragmenter[ollama]",
    "fragmenter[huggingface]",
]

[project.scripts]
fragmenter = "fragmenter.cli:main"

[project.urls]
Homepage = "https://github.com/RISE-Dependable-Transport-Systems/fragmenter"
Repository = "https://github.com/RISE-Dependable-Transport-Systems/fragmenter"
Issues = "https://github.com/RISE-Dependable-Transport-Systems/fragmenter/issues"
Changelog = "https://github.com/RISE-Dependable-Transport-Systems/fragmenter/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fragmenter"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src",
    "tests",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/fragmenter"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = ["loguru.*"]
ignore_missing_imports = true

[dependency-groups]
test = ["pytest>=8.0.0", "pytest-cov>=6.0.0", "pytest-mock>=3.14.0"]
dev = [
    "matplotlib>=3.10.7",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "ragas>=0.4.0",
    "ruff>=0.14.7",
    "rust-just>=1.45.0",
    "scikit-learn>=1.7.2",
    "umap-learn>=0.5.9.post2",
]
